+++ app/components/index.tsx
@@ -1,0 +1,7 @@
+import Button from "./Button";
+import FormField from "./FormField";
+import PageHeading from "./PageHeading";
+import Spinner from "./Spinner";
+import TableTop from "./Tabletop";
+
+export { Button, FormField, PageHeading, Spinner, TableTop };
